chore: use conditional alert popover in timeline table and import the scss file
